Some icky string munging
So, since we're reading bytes from the process descriptor, sometimes the
read will split in the MIDDLE of a UTF8 wide character. This makes the
resulting string invalid, meaning that any string processing on it will
raise an exception. We can't just drop invalid characters, because that
means that random high characters all over get missed.

Instead, I keep a `carry` array of bytes. I find the last low bit in the
command output string, since that's the last one I *know* is valid and
split the string right there. The trailing bytes are saved in `carry`
and the rest is re-encoded as a UTF8 string and pumped back to the
client.

I'd much rather just stuff it into the websocket and let the hterm
client reassemble it, but apparently the `ws.send()` does some string
processing and that barfs on the invalid bytes.