Add a ttl setting
When a ttl is set, the session won't die as soon as the websocket
connection closes. Instead, it will give you a grace period. If you
reconnect within that grace period, your session will be resumed
without interruption--even if you were running applications at the time.

This is currently not production ready. There are some quirks I want to
iron out, such as the fact that if you're running Vim, it forgets how to
use the arrow keys temporarily when you reconnect (!!). I'm also not
conviced that the cache invalidation is working perfectly yet.

This will be documented once these issues have been corrected and it's
ready for release.