Can now almost fully restore a session
For example, if you're running Vim, or the like, the terminal state will
mostly be restored. The part that won't is the secondary buffer, so when
you exit vim, you'll have some screen garbage and you'll need to clear
or reset the terminal. I'm not sure there's really a good way to avoid
that, so I'm going to punt on it for the time being.

The other limitation is that with `ttl` enabled, you can only have a
single session active with any given browser. If you start a second
session, the first one will be terminated, though you won't actually
lose any state.